// src/app/modules/shared/components/data-table/models/options.interface.ts
export interface TableOptionsInterface {
  add?: boolean;
  search?: boolean;
  import?: boolean;
  export?: boolean;
  check?: boolean;
  loading?: boolean;
  reorder?: boolean;
  details?: boolean;
  VisibleColumn?: boolean;
  columnSearch?: boolean;
  globalSearch?: boolean;
  custom?: string;
  editable?: boolean;
}